requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
tzlocal==4.2
whitenoise==6.2.0
requests